json read added

+++ src/com/company/Classes/Clouds.java
@@ -4,16 +4,16 @@ import java.io.Serializable;
 
 public class Clouds implements Serializable {
 
-    public Clouds(Integer all) {
+    public Clouds(Long all) {
         this.all = all;
     }
 
-    private Integer all;
+    private Long all;
 
-    public Integer getAll() {
+    public Long getAll() {
         return all;
     }
-    public void setAll(Integer all) {
+    public void setAll(Long all) {
         this.all = all;
     }
